back/ai_test_platform/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

python-decouple==3.8

psycopg2-binary==2.9.10

openai==1.98.0

requests==2.32.4

asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
